Fixed encoding issues (closes #71)
The root cause was a weird char encoding in a authorship comment in
Rubygems' rubygems/package/tar_input.rb

Refactored the processing of skipped lines in simplecov to be more in line
with the other facilities in the library, specifically to use SimpleCov::SourceFile::Line
instead of raw source arrays.

Due to this, the processing of skipped lines in results only happens on the filtered
resultset, which won't include any gem/library files. This should also make result generation
faster.