Recurse upward looking for .simplecov
The SimpleCov documentation states: "To avoid [duplicated configuration], you
can place a file called .simplecov in your project root." However, SimpleCov did
not actually look in the "project root" for this file, unless specially configured
to know where the project root is. Rather, it would simply look in the working
directory. Working directory != project root. (Unfortunately.)

Therefore, when looking for .simplecov, recurse upwards from the working directory
until it is found (or the root directory is reached). This is comparable to the
way that git recurses upwards looking for a .git directory, meaning that you can
use git anywhere in your project root or below.

Since there is no guarantee that you have permissions to read .simplecov, catch
exceptions and display a warning message if one occurs.