Change nil precedence when merging result sets
When merging two test suites, it might be the case that some
files are loaded by one suite and not loaded by the other one.

Whenever this happens, the "never" lines from the test suite that
loaded the file should trump the "uncovered" lines from the test
suite that didn't.

This commit solves that by giving higher precedence to `nil` values
when merging.