Make `track_files` work when configured with nil
Fixes #462.

The implementation of `track_files` was refactored recently to suppress
a warning in the underlying instance variable had not been initialized.
This had the effect of changing the behavior when trying to nullify the
default configuration of tracked files in some cases through the use of
nil.

With this change, a declaration of `track_files nil` will once again
clear out any previous configuration and essentially disable the
feature.

Reference:
  * https://github.com/colszowka/simplecov/pull/447
  * https://github.com/colszowka/simplecov/issues/462