Test current JRuby release
I thought it'd be worth running against the current JRuby release, not just head. I didn't want to interfere with a support decision (although I'd love to see JRuby officially supported!) so added it to the allow_failures for now. Thanks for the great work, long time simplecov user :) (and with that also the whole shoes4 project on JRuby) !