Officially Support JRuby 9.1+
I've personally been running simplecov with JRuby 1.7+ for years
so I'd say it is good to go either way ;P

The reality of edge cases is more complicated but it seems the
fine folks at @jruby (namely @enebo ) made the last fixes
necessary as seen in jruby/jruby#1196 so that they're available
in 9.1+ releases! The others should work, but have some little
failures (9.0.5.0 fails on one cuke for me).

So, the easiest way forward is to support 9.1+. Sadly the tests
take a while as they always start new interpreter instances and
startup time isn't exactly JRuby's strength, but CI does it
so it's bearable. :)

If anyone has problems with JRuby on a PR, in an issue or whatever
please feel free to ping me (@PragTob)

Also, special thanks goes to @donv who created did a lot of testing,
issue reporting et. al. for simplecov on JRuby for a long time!

* fixes #524