Rewrite merge helpers to work with frozen objects
Coverage.result returns a frozen Hash containing frozen Strings as keys
and frozen Arrays as values. The existing merge helper did not work when
passed a Coverage.result directly, because it would try to extend the
frozen Arrays. This is not a problem in typical SimpleCov usage because
usually Coverage.result has been roundtripped through .resultset.json
resulting in an unfrozen copy. But when trying to do manual merging of
Coverage.result hashes the use of extend gets in the way.

The merge helper has now been rewritten a method that takes two Hashes
and doesn't modify them or their contents at all. While I was touching
this code I tried to clean it up a little bit as well.