Stop handling string filters as regexps
Previously, filter strings passed to `add_filter` have been handled as regexps.
It means, for example, `add_filter '.pl'` matches against 'sample.rb'
because `.` is handled as _any character_ in regexp.

As now we have [regexp filter](https://github.com/colszowka/simplecov/pull/589),
there's no reason to handle filter strings as regexps.

However we need to think about semver with this change.
If this behavior was intentional,
we cannot introduce this change until next major version bump.
Or we can do if this was a _bug_.