Break default profiles out into new files
The goal behind this PR is to simplify the ```defaults.rb``` file. It
just breaks out the default profiles into their own files to
reduce the noise in this ```defaults.rb``` file

Things I like about this PR:

- It simplifies ```defaults.rb```
- It gives structure to any more default profiles that will be added, or
other default files to be added.

Things I don't like about this PR:

-Typically folders are used for namespacing and the ```profiles/ folder
is not used in that way in the PR.
- There will probably be very few if any more default profiles added, so
point 2 above has little weight.