Update alternate-formatters.md with simplecov-erb
I needed to generate output from simplecov that displayed in the output of my CI jobs. I implemented this via a user-configurable ERB template, so that the output format can be customized with minimal friction. Source is at https://github.com/kpaulisse/simplecov-erb and it's available under the MIT license.