Performance: memomize LinesClassifier::no_cov_line
In my measurements the majority of the time in SimpleCov was being
spent in this one-line function.

In my tests on a large project, this change makes SimpleCov run
from 2.5x to 3.75x faster.

Admittedly, memoizing it does change its behavior but only in the edge
case where someone runs SimpleCov twice in a single test run, changing
the +nocov_token+ between the two runs. This is such an odd edge case I
decided not to worry about it.