Don't crash on invalid UTF-8 byte sequences.
If a line of code has an invalid byte sequence in UTF-8, count it
as a relevant line rather than crashing.

Useful for projects that use `track_files "**/*.rb"` and have the
builder gem in a subdir such as vendor. Builder has a line of code with
an invalid UTF-8 byte sequence.