Fix problem with coverage merging
I was seeing the LOC in the merged set coming out less than the LOC in
one of the input result sets. That seems illogical. The merged result
LOC should be greater than or equal to the LOC in any given input result
set.

Tracked it down to the merge logic in the case that the line has 0
hits.

The logic was:

  if this line was never loaded in input 1 or never loaded in input 2
    it was never loaded
  else
    sum the coverage
  end

The problem with this is that the line was loaded and missed in one of
the input result sets. The merged result set considers this line "never"
and one of the inputs considered it "missed". In calculating LOC, missed
lines are included and never lines are not included. Hence this line
would be included in the input LOC count but not the merged result LOC
count.

I think the logic should be:

  if this line was never loaded in input 1 or input 2
    it was never loaded
  else
    sum the coverage
  end

Then the 0 value will propagate to the merged result, causing
this line to be included in LOC in the merged result.