Avoid hanging with parallel_tests
ParallelTests.last_process? retuns true if the process is the last one to start.  We are
interested in the last test to finish so we want to use ParallelTests.number_of_running_processes

Because it's possible to have the parallel_tests gem in your project and still run tests serially we
want to also check if tests are currently being run in parallel.  If ENV['TEST_ENV_NUMBER'] is defined
we know tests are currently being run in parallel.