Refactor combiners to be more functional
Essentially all they do is combine results - no need for state,
just take on result, combine it with another and return the
result.

Makes code a bit simpler and also removes the need for memory
allocation.

Not sure how people like me passing module names to the `Combine`
module, but this allows me to extract the concern there that
it checks for existing coverage so it doesn't have to be repeated.