(Hopefully) fix #776 by returning empty responses on not found file
Couldn't reproduce locally.
As the failure was only ever JRuby, one test and the gif at that
(which shouldn't matter for the result of the test as a whole)
decided to just roll with it with an empty response and log
when it happens so that we could track other failures back to it.