Change wording of branch coverage to coverage_criterion (#793)
This is more flexible as it also accommodates for all the other
coverage types. It would also allow us to use an array like
`[:branch, :lines, :methods]` or something in the future.

It's also arguable a better name, let me know if you disagree :)

I'm also half-way sure that this will intermittently break
branch coverage but that's why adding an integration test there
is next on the list!