Implement skipping lines in a way that's friendlier for branches
The previous implementation was very focussed on lines (surprise!)
and shoe horning branches in there would have been difficult.

Plus we probably (in the future) want to be able to run only
branch coverage so we needed a completely separate way to run
them anyhow.

This introduces a new structure of "no_cov_chunks" that can then
be used to work with every type of coverage to see if there are
intersections or what not to skip those coverages.