Fix crash when running against old version of file
We do a light check on every result if it looks like it conforms
to our file format. If not, we ignore it and throw it out.

Initial plan was to just catch an error and return a known good
result, but sadly since we don't know what order the results come
in that doesn't really work (as the first could be the "bad" result).
Might be able to change this if we always knew we started with
a known good result.

Checking feels kinda clunky, but I honestly didn't know of
a better way right now.