Stop generally symbolizing all the keys
A lot of them we convered back to strings anyhow, plus with
frozen string literal the gains are negligible anyhow.

Also, always symbolizing everything in the JSON.parse was a bit
of an overkill especially when you imagine a code base with lots
of files and lots of branches we could easily contribute to
blowing up memory considerably/beyond an accepatable threshhold.