Let the HTML report show the branch type instead of +/-
I believe this is much more user friendly as opposed to the
thinking what is positive/negative (like in an unless ... else,
or case when else). This way people see immediately that the else
branch was/wasn't hit or the then branch or the body of a loop.

Specifically I think it also helps with us reporting else branches
that aren't even defined (for if/case) - when people see 0 hits
for an else branch I think that's easier to figure out than
0 hits for the "-" branch. There's less of a "what does -/+ mean?"
going on.

As a side effect, also removes some code for checking what's
positive that's also rather intricately tied to how branch
coverage hands out its IDs which I don't believe we should take
for granted forever.