Adapt old resultset.jsons instead of just warning
After discussion with @deivid-rodrigues I realized that adapting
the result isn't actually that difficult and it's actually
a better solution to my current problem (that the warning
gets printed multiple times).

Yes, one could have moved the code around so that wrong format
results wouldn't even be written back, but Deivid was right
it might be useful for some people to keep the result and it's
actually not more complicated than sorting them out.

So this change fixes the multiple warnings as well as keeping
old results around.

Also moved the checking/changing to a fairly central place to
safeguard against old results flying in from somewhere else.