Remove covered_strength rounding
It was the only non output place in the entire code base where
we rounded it as best as I can tell. Rounding can be done in
formatters (imo) and by doing it on the source file level
we lost precision in the accumulation for all files.

Went back all the way to #59 and couldn't see a reason why
it _had_ to be rounded. Also, no tests fail so :shrug: