Use CoverageData everywhere for consistency
This allows us to get rid off duplicated and inconsistent
calculations. As a side effect of this, #565 is fixed now.

So, fixes #565.

The bigger purpose is multi step though where the access
to coverage data is normalized and hence methods that work
with expecting a minimum amount of coverage and formatters
can work with the known good same data base as opposed
to memoizing all different method names.

At first I wanted to only do it for FileList and leave the
rest for later. But the strength computation needed the
total number of lines of code and so to pass it in I'd have
had to calculate it but I wanted to calculate it only in
CoverageData and there the avalanche started....

Commit also includes moving methods around in SourceFile to be
private - it happened in the moment, forgive me for no extra
commit (it was hard to determine what I could delete/what I need
to adjust etc.).