Amend 0.18 changelog with notice about code climate test reporter
I didn't add this information as to my mind we didn't change any API documented anywhere and hence it shouldn't appear in a Changelog.

However, this affects people and they have to find out for themselves so notifying them of this might be a good thing. I still couldn't bring myself to put it under breaking changes as technically we haven't broken anything that we ever communicated as working.

So a link to https://github.com/codeclimate/test-reporter/issues/413 seems to be helpful.