Only run our minitest plugin code if SimpleCov is indeed defined
* Fixes #877
* not quire sure whether we should wrap the entire thing in
if undefined?(..) or not but since the file has all the right
structure otherwise I'm more worried about minitest being
weirded out that it doesn't find what it expects than I am
of letting it execute an empty method
* also yes the version workaround is maybe a bit weird, however
I wanted our test suite to reproduce the exact failure case
because otherwise it'd be to easy to just implement a
respond_to? check which would have everything passing but would
fail in real life.