Refactor result_exit_status
This method was giant spaghetti, that was hard to maintain and
fix stuff in.

The new solution might seem overblown but with it's current
interdependencies was the best I could come up with.

Decided to decouple it a bit from asking SimpleCov itself for
the coverage values to make it easier to test in isolation. No
tests added as of right now though since it's tested quite a bit.

The strategy to put each check into its own class produces more
lines of code but makes their intent clearer and also makes it
easier to figure out what belongs to what check.