Disentangle exit status handling & differntiate our vs. their err
Refactoring this part made it clearer that all the different
checks against success weren't necessary and could be done on
a higher level removing lots and lots of branches on the lower
level all checking for the same thing.

Also implements the feature that we print different error
messages for aborting on "we want to throw an error" vs.
"we determined there was an error before".

Decided to print out both in case we have a false positive then
people can file better error reports of "you said there was
an error before, but there wasn't!"

Also deleted some tests as they dealt with cases that don't
happen anymore/don't have to be checked anymore. Added some cukes
to show that we say something else went wrong.

Test suit holds up. Let's hope this won't be another billion
fixes for minute details I forgot about and need to fix after
the next release!