Remove markdown content to limit gem size
We don't need to ship the entire history of the gem, so split
into current and old (somewhat arbitrarily).
Similarly, CONTRIBUTING, CODE_OF_CONDUCT and ISSUE_TEMPLATE are
mostly relevant for contributing to simplecov and don't need to
ship with the gem.

Not too bad for ~20% space savings:

```
tobi@speedy:~/github/simplecov(remove-markdown-to-limit-gem-size)$ ls -l pkg/
total 104
-rw-r--r-- 1 tobi tobi 45568 Nov 29 22:04 simplecov-0.20.0.gem
-rw-r--r-- 1 tobi tobi 56832 Nov 29 21:57 simplecov-0.20.0.gem.old
```