Fix minimum_coverage_by_file check & prep 0.21.1
minimum_coverage_by_file was a gaping hole in the test suite,
which is now partially remedied through rather thorough cucumber
scenarios and 2 small unit tests that should be expanded on in
the future.

On the negative side: This will inconvenince some people. On the
positive side stuffed another hole in the test suite towards a
better future :D Also it was reported very soon thanks to @iMacTia

Fixes #965