Add needed require line for CSV example
We got confused when trying to use the CSVFormatter and discovered it also required another gem (at least for our version). Thus I suggest adding a clarification as I suggest in this PR. It might also be more clear if the "Multiple formatters" heading was below the "JSON formatter" section.