Reduce GC pressure in LinesCombiner by ~90%
The LinesCombiner uses the `Array#zip` method to merge both line
coverage arrays A and B together. This creates a new array, immediately
discarding A and B.

When we're combining a lot of files, this creates significant pressure
on the garbage collection because we're allocating a lot of intermediate
arrays.

For example, all GitLab unit tests emit about 328 MB of minified,
uncompressed `.resultset.json` files. If we use `SimpleCov.collate` to
merge them (using the default configuration), we invoke the GC over
1,700 times.

If we instead reuse the longest array A or B to merge coverage in-place
without creating a new array, we reduce the GC invocations to 55.

This results in a total reduction in GC time by around 90%. On my M3
Max, this reduces the garbage collection time with our unit test
coverage result sets from 2.78s to 0.32s. In total, we're observing a
total speed up of coverage merging in this example of about 25% (9.55 ->
7.05s).

Because both approaches of using `Array#zip` or updating the array
in-place are functionally identical, we can replace the former with the
latter without any other changes.